Relocate disconnect cluster functionality to external cluster service for separation of concerns
