+++ src/app/cluster/list/external-cluster/component.ts
@@ -39,6 +39,7 @@ import {
   ExternalClusterState,
 } from '@shared/entity/external-cluster';
 import {ExternalClusterDeleteConfirmationComponent} from '@app/cluster/details/external-cluster/external-cluster-delete-confirmation/component';
+import {ExternalClusterService} from '@core/services/external-cluster';
 
 @Component({
   selector: 'km-external-cluster-list',
@@ -63,6 +64,7 @@ export class ExternalClusterListComponent implements OnInit, OnChanges, OnDestro
 
   constructor(
     private readonly _clusterService: ClusterService,
+    private readonly _externalClusterService: ExternalClusterService,
     private readonly _projectService: ProjectService,
     private readonly _userService: UserService,
     private readonly _router: Router,
@@ -186,8 +188,7 @@ export class ExternalClusterListComponent implements OnInit, OnChanges, OnDestro
 
   disconnectClusterDialog(cluster: ExternalCluster, event: Event): void {
     event.stopPropagation();
-
-    this._clusterService.showDisconnectClusterDialog(cluster, this._selectedProject.id);
+    this._externalClusterService.showDisconnectClusterDialog(cluster, this._selectedProject.id);
   }
 
   deleteClusterDialog(cluster: ExternalCluster, event: Event): void {
